move lib.rs harness into harness.rs
